Move halo code to Module:MatchGroup/Input/Custom
